move packages to util folder
