+++ src/proxy/proxy.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 
 	"github.com/skycoin/teller/src/daemon"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 
 	"io"
 	"time"
