+++ src/service/exchange/exchange_test.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/skycoin/teller/src/dbutil"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/dbutil"
+	"github.com/skycoin/teller/src/util/logger"
 	"github.com/skycoin/teller/src/service/scanner"
 	"github.com/skycoin/teller/src/service/sender"
 	"github.com/skycoin/teller/src/service/testutil"
