+++ src/service/monitor/monitor.go
@@ -10,8 +10,8 @@ import (
 	"fmt"
 
 	"github.com/skycoin/teller/src/daemon"
-	"github.com/skycoin/teller/src/httputil"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/httputil"
+	"github.com/skycoin/teller/src/util/logger"
 	"github.com/skycoin/teller/src/service/exchange"
 )
