+++ src/service/service.go
@@ -9,7 +9,7 @@ import (
 	"io"
 
 	"github.com/skycoin/teller/src/daemon"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 )
 
 const (
